Switch component over to new structure
